Fix Android timestamp overflow that blocked all space config sync (#79)
The Android crypto module read message timestamps with optInt (32-bit).
Millisecond-epoch timestamps overflow and wrap negative; the JS batch path
matches results to their source by exact timestamp, so the match always
failed and every incoming control message (space-manifest config updates,
etc.) was silently dropped. Read with optLong (64-bit) instead. iOS and the
JS/desktop path were already correct - Android only.

Also refresh the channels/spaces queries when a manifest is applied (so UI
updates immediately) and add debug-level receive-path logging. Mobile only.
Runtime-verified on a physical Android device.

+++ modules/quorum-crypto/android/src/main/java/expo/modules/quorumcrypto/QuorumCryptoModule.kt
@@ -880,7 +880,14 @@ class QuorumCryptoModule : Module() {
 
                         for (m in 0 until messagesArr.length()) {
                             val msg = messagesArr.getJSONObject(m)
-                            val timestamp = msg.optInt("timestamp", 0)
+                            // Use optLong (64-bit): timestamps are millisecond-epoch
+                            // values (~1.7e12) that overflow a 32-bit Int, wrapping to
+                            // negative. The overflowed value is echoed into every result
+                            // and later matched against the original message's full-
+                            // precision timestamp in JS (WebSocketContext batch path),
+                            // so optInt silently dropped all control messages (e.g.
+                            // space-manifest updates) on Android.
+                            val timestamp = msg.optLong("timestamp", 0L)
 
                             val ephemeralPubKeyHex = msg.optString("ephemeral_public_key", "")
                             val envelope = msg.optString("envelope", "")
@@ -1146,7 +1153,14 @@ class QuorumCryptoModule : Module() {
 
                         for (m in 0 until messagesArr.length()) {
                             val msg = messagesArr.getJSONObject(m)
-                            val timestamp = msg.optInt("timestamp", 0)
+                            // Use optLong (64-bit): timestamps are millisecond-epoch
+                            // values (~1.7e12) that overflow a 32-bit Int, wrapping to
+                            // negative. The overflowed value is echoed into every result
+                            // and later matched against the original message's full-
+                            // precision timestamp in JS (WebSocketContext batch path),
+                            // so optInt silently dropped all control messages (e.g.
+                            // space-manifest updates) on Android.
+                            val timestamp = msg.optLong("timestamp", 0L)
                             val isDREnvelope = msg.optBoolean("is_double_ratchet_envelope", false)
                             val isInitEnvelope = msg.optBoolean("is_init_envelope", false)
                             val encryptedContent = msg.optString("encrypted_content", "")
